Move macros from config to detail/config
