+++ include/boost/charconv/detail/integer_search_trees.hpp
@@ -8,7 +8,7 @@
 // https://stackoverflow.com/questions/1489830/efficient-way-to-determine-number-of-digits-in-an-integer?page=1&tab=scoredesc#tab-top
 // https://graphics.stanford.edu/~seander/bithacks.html
 
-#include "../config.hpp"
+#include <boost/charconv/detail/config.hpp>
 #include <limits>
 #include <array>
 #include <cstdint>
