+++ include/boost/charconv/detail/memcpy.hpp
@@ -5,6 +5,7 @@
 #ifndef BOOST_CHARCONV_DETAIL_MEMCPY_HPP
 #define BOOST_CHARCONV_DETAIL_MEMCPY_HPP
 
+#include <boost/charconv/detail/config.hpp>
 #include <boost/charconv/detail/is_constant_evaluated.hpp>
 #include <cstring>
 #include <cstdint>
